Fixes from Dick Franks:
Perl 5.6.2 throws an unexpected warning about subroutine profiles for
SEQUENTIAL() and friends in DNS.pm.
Perl 5.8.3 and earlier throws a warning for Domain.pm if declared
constants are undefined. This prevents the unwanted bits of code from
being optimised away.
DomainName.pm modified to make it compatible with existing dn_comp().
01-resolver.t modified to display Cygwin or Win32 as appropriate. At
present there appears to be no way to discover which configuration is
being used on Cygwin platform.


git-svn-id: https://www.net-dns.org/svn/net-dns/trunk@964 a29f3fcb-687d-e111-8194-001676cd6a66